+++ np.Templating/CHANGELOG.md
@@ -4,6 +4,11 @@
 
 See Plugin [Documentation](https://noteplan.co/templates/docs) for details on available commands and use case.
 
+## [2.0.13] 2025-08-05 @dwertheimer
+- Fix bug where inline H1 title was not being used in templateNew (thx @crussell)
+- Ensure that inline H1 title is not created in frontmatter even if there is other frontmatter being created
+- if there is newNoteTitle and also an inline H1 title, the newNoteTitle will take precedence and will be created in frontmatter
+
 ## [2.0.12] 2025-08-02 @dwertheimer
 - Fix templateNew to handle blank meeting note edge case